Move functions to a separate file
Part of the refactoring work, move common functions to bot_utils.py so
they can be shared by various bot commands and used in the internal
implementation of the bot.
