+++ slackbot/bot_utils.py
@@ -1,0 +1,27 @@
+import string
+import slack_sdk as slack
+from bot_info import slack_token, botid, workerid
+
+def extract_command(msg):
+    cmd = msg["text"].replace(workerid, "").replace(botid, "")
+    cmd = cmd.translate(str.maketrans('', '', string.punctuation))
+    cmd = cmd.lower()
+    return "".join(cmd.split())
+
+
+def replyto(msg, reply, username=workerid, broadcast=False):
+    sc = slack.WebClient(slack_token, timeout=300)
+    channel = msg['channel']
+    userid = msg['user']
+    thread_ts = msg['thread_ts'] if 'thread_ts' in msg else msg['ts']
+    reply_msg = f"<@{userid}> {reply}"
+    rc = sc.chat_postMessage(
+        username=username,
+        channel=channel,
+        thread_ts=thread_ts,
+        reply_broadcast=broadcast,
+        text=reply_msg
+    )
+    if not rc["ok"]:
+        print("Failed to send slack message")
+        print(rc)
